backend/requirements.txt
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
python-dotenv==1.0.1
APScheduler==3.10.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
pytest==8.0.2
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
isort==5.13.2
supabase==1.2.0
requests==2.31.0
python-dateutil==2.8.2
PyJWT==2.8.0
httpx==0.24.1
postgrest==0.10.8
